Refactor: move parallel_grid from src_parallel to module_hamilt_pw/hamilt_pwdft/
