+++ source/module_elecstate/module_charge/charge.cpp
@@ -23,7 +23,7 @@
 #include "module_hamilt_pw/hamilt_pwdft/global.h"
 #include "charge.h"
 #include "module_elecstate/magnetism.h"
-#include "src_parallel/parallel_grid.h"
+#include "module_hamilt_pw/hamilt_pwdft/parallel_grid.h"
 #include "module_base/math_integral.h"
 #include "module_base/math_sphbes.h"
 #include <vector>
