+++ source/module_hamilt_pw/hamilt_pwdft/global.h
@@ -298,7 +298,7 @@ extern pseudopot_cell_vnl ppcell;
 
 #include "module_cell/unitcell.h"
 #include "module_cell/module_symmetry/symmetry.h"
-#include "src_parallel/parallel_grid.h"
+#include "module_hamilt_pw/hamilt_pwdft/parallel_grid.h"
 #include "src_parallel/parallel_kpoints.h"
 namespace GlobalC
 {
